requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
torchvision
httpx
opencv-python
pillow
